backend/app/core/security.py
from datetime import datetime, timedelta
from typing import Optional, Any, Union
from jose import jwt
from app.core.config import settings
import hashlib
import bcrypt

def create_access_token(subject: Union[str, Any], expires_delta: Optional[timedelta] = None) -> str:
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt

def verify_password(plain_password: str, hashed_password: str) -> bool:
    # Pre-hash password to handle length limit of bcrypt (consistent with get_password_hash)
    hashed_plain = hashlib.sha256(plain_password.encode('utf-8')).hexdigest()
    
    # bcrypt.checkpw expects bytes. hashed_password from DB is str.
    return bcrypt.checkpw(
        hashed_plain.encode('utf-8'), 
        hashed_password.encode('utf-8')
    )

def get_password_hash(password: str) -> str:
    # Pre-hash password with SHA256 to handle 72-byte bcrypt limit
    hashed_plain = hashlib.sha256(password.encode('utf-8')).hexdigest()
    
    # bcrypt.hashpw returns bytes, we decode to str for storage
    pwd_bytes = hashed_plain.encode('utf-8')
    salt = bcrypt.gensalt()
    return bcrypt.hashpw(pwd_bytes, salt).decode('utf-8')
