backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
duckdb>=0.9.2
duckdb-engine>=0.9.2
sqlalchemy>=2.0.25
asyncpg>=0.29.0  # Kept for potential Postgres move, though DuckDB handles local
websockets>=12.0
pytest>=7.4.0
httpx>=0.26.0
pydantic[email]
